// src/app/@core/api/model/constatDto.ts
/**
 * EPR api
 * The list of EPR project endpoints
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ProjetDto } from '././projetDto';
import { RiskDto } from '././riskDto';


export interface ConstatDto { 
    id: string;
    date: string;
    time: string;
    semaine: number;
    typeRisk: number;
    risk: RiskDto;
    zone: string;
    type: ConstatDto.TypeEnum;
    rank: ConstatDto.RankEnum;
    company: string;
    contact: string;
    local: string;
    site: string;
    action: string;
    description: string;
    picture: string;
    projet: ProjetDto;
    faitMarquant: boolean;
    finished: boolean;
}
export namespace ConstatDto {
    export type TypeEnum = 'log' | 'risk';
    export const TypeEnum = {
        Log: 'log' as TypeEnum,
        Risk: 'risk' as TypeEnum
    };
    export type RankEnum = 'positive' | 'negative';
    export const RankEnum = {
        Positive: 'positive' as RankEnum,
        Negative: 'negative' as RankEnum
    };
}